third_party/blink/web_tests/external/wpt/html/semantics/popups/popup-invoking-attribute.tentative.html

<!DOCTYPE html>
<meta charset="utf-8" />
<title>Popup invoking attribute</title>
<link rel="author" href="mailto:masonf@chromium.org">
<link rel=help href="https://open-ui.org/components/popup.research.explainer">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<button togglepopup=p1>Toggle Popup 1</button>
<div popup=popup id=p1>This is popup #1</div>

<style>
  [popup] {
    border: 5px solid red;
    top: 100px;
    left: 100px;
  }
</style>

<script>
function clickOn(element) {
  const actions = new test_driver.Actions();
  return actions.pointerMove(0, 0, {origin: element})
    .pointerDown({button: actions.ButtonType.LEFT})
    .pointerUp({button: actions.ButtonType.LEFT})
    .send();
}

const popup = document.querySelector('[popup]');
const button = document.querySelector('button');
let showCount = 0;
let hideCount = 0;
popup.addEventListener('show',() => ++showCount);
popup.addEventListener('hide',() => ++hideCount);

async function assertState(expectOpen,expectShow,expectHide) {
  assert_equals(popup.matches(':popup-open'),expectOpen,'Popup open state is incorrect');
  await new Promise(resolve => requestAnimationFrame(resolve));
  assert_equals(showCount,expectShow,'Show count is incorrect');
  assert_equals(hideCount,expectHide,'Hide count is incorrect');
}

promise_test(async () => {
  showCount = hideCount = 0;
  await assertState(false,0,0);
  await clickOn(button);
  await assertState(true,1,0);
  popup.hidePopup();
  await assertState(false,1,1);
  button.click();
  await assertState(true,2,1);
  popup.hidePopup();
  await assertState(false,2,2);
}, "Clicking a togglepopup button opens a closed popup");

promise_test(async () => {
  showCount = hideCount = 0;
  await assertState(false,0,0);
  await clickOn(button);
  await assertState(true,1,0);
  await clickOn(button);
  await assertState(false,1,1);
}, "Clicking a togglepopup button closes an open popup");

["popup","hint","async"].forEach(type => {
  [0,1,2].forEach(t => {
    [0,1,2].forEach(s => {
      [0,1,2].forEach(h => {
        const popup1 = Object.assign(document.createElement('div'),{popup: type, id: 'popup-1'});
        const popup2 = Object.assign(document.createElement('div'),{popup: type, id: 'popup-2'});
        assert_not_equals(popup1.id,popup2.id);
        assert_true(!document.getElementById(popup1.id));
        assert_true(!document.getElementById(popup2.id));
        const button = document.createElement('button');
        document.body.appendChild(popup1);
        document.body.appendChild(popup2);
        document.body.appendChild(button);
        if (t) button.setAttribute('togglepopup',t===1 ? popup1.id : popup2.id);
        if (s) button.setAttribute('showpopup',s===1 ? popup1.id : popup2.id);
        if (h) button.setAttribute('hidepopup',h===1 ? popup1.id : popup2.id);
        test(() => {
          // This mimics the expected logic:
          let expectedBehavior = t ? "toggle" : (s ? "show" : (h ? "hide" : "none"));
          let expectedId = t || s || h || 1;
          if (!t && s && h) {
            // Special case - only use toggle if the show/hide idrefs match.
            expectedBehavior = (s === h) ? "toggle" : "show";
          }
          const otherId = expectedId !== 1 ? 1 : 2;
          function assert_popup(num,state,message) {
            assert_true(num>0);
            assert_equals((num===1 ? popup1 : popup2).matches(':popup-open'),state,message || "");
          }
          assert_popup(expectedId,false);
          assert_popup(otherId,false);
          button.click();
          assert_popup(otherId,false,'The other popup should never change');
          switch (expectedBehavior) {
            case "toggle":
            case "show":
              assert_popup(expectedId,true,'Toggle or show should show the popup');
              (expectedId===1 ? popup1 : popup2).hidePopup(); // Hide the popup
              break;
            case "hide":
            case "none":
              assert_popup(expectedId,false,'Hide or none should leave the popup hidden');
              break;
            default:
              assert_unreached();
          }
          (expectedId===1 ? popup1 : popup2).showPopup(); // Show the popup
          assert_popup(expectedId,true);
          assert_popup(otherId,false);
          button.click();
          assert_popup(otherId,false,'The other popup should never change');
          switch (expectedBehavior) {
            case "toggle":
            case "hide":
              assert_popup(expectedId,false,'Toggle or hide should hide the popup');
              break;
            case "show":
            case "none":
              assert_popup(expectedId,true,'Show or none should leave the popup showing');
              break;
            default:
              assert_unreached();
          }
        },`Test ${button.outerHTML} with popup=${type}`);
        button.remove();
        popup1.remove();
        popup2.remove();
      });
    });
  });
});
</script>
